+++ samples/samples-aspnet/embedded-auth-with-sdk/embedded-auth-with-sdk/Controllers/ManageController.cs
@@ -53,7 +53,7 @@ namespace embedded_auth_with_sdk.Controllers
             {
                 NewPassword = model.NewPassword,
             };
-            
+
             try
             {
                 var authnResponse = await _idxClient.ChangePasswordAsync(changePasswordOptions, (IIdxContext)Session["idxContext"]).ConfigureAwait(false);
@@ -79,7 +79,7 @@ namespace embedded_auth_with_sdk.Controllers
             }
             catch (OktaApiException exception)
             {
-                ModelState.AddModelError("Oops! Something went wrong.", exception.ErrorSummary 
+                ModelState.AddModelError("Oops! Something went wrong.", exception.ErrorSummary
                                          ?? "Cannot change password. Check if the new password meets the requirements.");
                 return View("ChangePassword", model);
             }
@@ -217,7 +217,7 @@ namespace embedded_auth_with_sdk.Controllers
             {
                 return View(view, model);
             }
-            
+
             var verifyAuthenticatorOptions = new VerifyAuthenticatorOptions
             {
                 Code = code,
@@ -244,7 +244,7 @@ namespace embedded_auth_with_sdk.Controllers
                         _authenticationManager.SignIn(new AuthenticationProperties(), identity);
                         return RedirectToAction("Index", "Home");
                 }
-                
+
                 return View(view, model);
             }
             catch (OktaException exception)
@@ -308,7 +308,7 @@ namespace embedded_auth_with_sdk.Controllers
                 Session["IdxContext"] = enrollResponse.IdxContext;
 
                 if (enrollResponse.AuthenticationStatus == AuthenticationStatus.AwaitingAuthenticatorVerification)
-                {   
+                {
                     return RedirectToAction("VerifyAuthenticator", "Manage");
                 }
 
@@ -327,17 +327,38 @@ namespace embedded_auth_with_sdk.Controllers
             var methodTypes = (List<AuthenticatorMethodType>)Session["methodTypes"]
                               ?? new List<AuthenticatorMethodType>();
             var model = new EnrollPhoneViewModel
-                            {
-                                MethodTypes = methodTypes,
-                                MethodType = methodTypes.FirstOrDefault(),
-                            };
+            {
+                MethodTypes = methodTypes,
+                MethodType = methodTypes.FirstOrDefault(),
+            };
 
             return View(model);
         }
 
+        public ActionResult EnrollGoogleAuthenticator()
+        {
+            var QrCode = (IQrCode)Session["qrcode"];
+            var sharedSecret = (string)Session["sharedSecret"];
+            var model = new EnrollGoogleAuthenticatorViewModel
+            {
+                QrCodeHref = QrCode.Href,
+                SharedSecret = sharedSecret,
+            };
+
+            return View(model);
+        }
+
+        [HttpPost]
+        [AllowAnonymous]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> EnrollGoogleAuthenticatorAsync(EnrollGoogleAuthenticatorViewModel model)
+        {
+            return View("VerifyTotp");
+        }
+
         public ActionResult SelectPhoneChallengeMethod(SelectAuthenticatorMethodViewModel model)
         {
-           return View(model);
+            return View(model);
         }
 
         [HttpPost]
@@ -358,7 +379,7 @@ namespace embedded_auth_with_sdk.Controllers
                     EnrollmentId = model.EnrollmentId,
                     MethodType = model.MethodType,
                 };
-                
+
                 var challengeResponse = await _idxClient.ChallengeAuthenticatorAsync(challengeOptions, (IIdxContext)Session["IdxContext"]);
 
                 switch (challengeResponse?.AuthenticationStatus)
@@ -375,22 +396,68 @@ namespace embedded_auth_with_sdk.Controllers
                 return View("SelectPhoneChallengeMethod", model);
             }
         }
-        
+
         public ActionResult SelectAuthenticator()
         {
             var authenticators = (IList<AuthenticatorViewModel>)Session["authenticators"] ?? new List<AuthenticatorViewModel>();
-            
+
             var viewModel = new SelectAuthenticatorViewModel
-                                {
-                                    Authenticators = authenticators,
-                                    AuthenticatorId = authenticators.FirstOrDefault()?.AuthenticatorId,
-                                    PasswordId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "password")?.AuthenticatorId,
-                                    PhoneId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "phone")?.AuthenticatorId,
-                                    WebAuthnId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "security key or biometric")?.AuthenticatorId,
-                                    CanSkip = TempData["canSkip"] != null && (bool)TempData["canSkip"]
-                                };
+            {
+                Authenticators = authenticators,
+                AuthenticatorId = authenticators.FirstOrDefault()?.AuthenticatorId,
+                PasswordId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "password")?.AuthenticatorId,
+                PhoneId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "phone")?.AuthenticatorId,
+                WebAuthnId = authenticators.FirstOrDefault(x => x.Name.ToLower() == "security key or biometric")?.AuthenticatorId,
+                CanSkip = TempData["canSkip"] != null && (bool)TempData["canSkip"]
+            };
 
             return View(viewModel);
+        }
+
+        [HttpPost]
+        [AllowAnonymous]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> VerifyTotpAsync(string code)
+        {
+            if (!ModelState.IsValid)
+            {
+                return View("VerifyTotp");
+            }
+
+            var verifyAuthenticatorOptions = new VerifyAuthenticatorOptions
+            {
+                Code = code,
+            };
+
+            try
+            {
+                var authnResponse = await _idxClient.VerifyAuthenticatorAsync(verifyAuthenticatorOptions, (IIdxContext)Session["idxContext"]);
+                Session["idxContext"] = authnResponse.IdxContext;
+
+                switch (authnResponse.AuthenticationStatus)
+                {
+                    case AuthenticationStatus.AwaitingPasswordReset:
+                        return RedirectToAction("ChangePassword", "Manage");
+
+                    case AuthenticationStatus.AwaitingAuthenticatorEnrollment:
+                        Session["authenticators"] = ViewModelHelper.ConvertToAuthenticatorViewModelList(authnResponse.Authenticators);
+                        TempData["canSkip"] = authnResponse.CanSkip;
+                        Session["isChallengeFlow"] = false;
+                        return RedirectToAction("SelectAuthenticator", "Manage");
+
+                    case AuthenticationStatus.Success:
+                        ClaimsIdentity identity = await AuthenticationHelper.GetIdentityFromTokenResponseAsync(_idxClient.Configuration, authnResponse.TokenInfo);
+                        _authenticationManager.SignIn(new AuthenticationProperties(), identity);
+                        return RedirectToAction("Index", "Home");
+                }
+
+                return View("VerifyTotp");
+            }
+            catch (OktaException exception)
+            {
+                ModelState.AddModelError(string.Empty, exception.Message);
+                return View("VerifyTotp");
+            }
         }
 
 
@@ -468,7 +535,7 @@ namespace embedded_auth_with_sdk.Controllers
 
                         selectAuthenticatorResponse = await _idxClient.SelectChallengeAuthenticatorAsync(selectAuthenticatorOptions, (IIdxContext)Session["IdxContext"]);
                     }
-                    
+
                     Session["IdxContext"] = selectAuthenticatorResponse.IdxContext;
 
                     switch (selectAuthenticatorResponse?.AuthenticationStatus)
@@ -483,6 +550,7 @@ namespace embedded_auth_with_sdk.Controllers
                                 MethodType = selectAuthenticatorResponse.CurrentAuthenticatorEnrollment.MethodTypes.FirstOrDefault(),
                             };
                             return View("SelectPhoneChallengeMethod", methodViewModel);
+
                         case AuthenticationStatus.AwaitingAuthenticatorVerification:
                             var action = (model.IsWebAuthnSelected)
                                 ? "VerifyWebAuthnAuthenticator"
@@ -512,22 +580,30 @@ namespace embedded_auth_with_sdk.Controllers
                     switch (enrollResponse?.AuthenticationStatus)
                     {
                         case AuthenticationStatus.AwaitingAuthenticatorVerification:
-                            if (model.IsPasswordSelected)
                             {
-                                return RedirectToAction("ChangePassword", "Manage");
-                            }
-                            else if (model.IsWebAuthnSelected)
-                            {
-                                Session["currentWebAuthnAuthenticator"] = enrollResponse.CurrentAuthenticator;
-                                return RedirectToAction("EnrollWebAuthnAuthenticator", "Manage");
-                            }
+                                if (model.IsPasswordSelected)
+                                {
+                                    return RedirectToAction("ChangePassword", "Manage");
+                                }
+                                // Just for consistency, can we use model.IsTotpSelected - Andrii do we need ....
+                                else if (enrollResponse.CurrentAuthenticator?.Name == "Google Authenticator")
+                                {
+                                    Session["qrcode"] = enrollResponse.CurrentAuthenticator.QrCode;
+                                    Session["sharedSecret"] = enrollResponse.CurrentAuthenticator.SharedSecret;
+                                    return RedirectToAction("EnrollGoogleAuthenticator", "Manage");
+                                }
+                                else if (model.IsWebAuthnSelected)
+                                {
+                                    Session["currentWebAuthnAuthenticator"] = enrollResponse.CurrentAuthenticator;
+                                    return RedirectToAction("EnrollWebAuthnAuthenticator", "Manage");
+                                }
 
-                            return RedirectToAction("VerifyAuthenticator", "Manage"); 
+                                return RedirectToAction("VerifyAuthenticator", "Manage");
+                            }
 
                         case AuthenticationStatus.AwaitingAuthenticatorEnrollmentData:
                             Session["methodTypes"] = enrollResponse.CurrentAuthenticator.MethodTypes;
                             return RedirectToAction("EnrollPhoneAuthenticator", "Manage");
-                            
 
                         default:
                             return View("SelectAuthenticator", model);
@@ -544,9 +620,9 @@ namespace embedded_auth_with_sdk.Controllers
         public ActionResult SelectRecoveryAuthenticator()
         {
             var viewModel = new SelectRecoveryAuthenticatorViewModel
-                                {
-                                    Authenticators = (List<AuthenticatorViewModel>)Session["authenticators"],
-                                };
+            {
+                Authenticators = (List<AuthenticatorViewModel>)Session["authenticators"],
+            };
             return View(viewModel);
         }